Update coordinate system
The confusion in the coordinate system comes
from the fact that in a point representation,
e.g. (x, y) y is the row and x is the column.
Whereas it when iterating over array it's
normal to think of (row, col), leading to
a mix-up between the graphics on-screen and
the in-game coordinates.