Small combat update
Player no longer damages themselves when
standing idle. When trying to move check
to see if there's an enemy on the target
tile and either attack or move there.

TODO: there's a lot of getters and setters,
instead make the base class's members protected
and modify those directly (?) from the derived
classes. And expose a proper narrower API
for everyone else.

TODO: an 'Action' class, instead of having a
ton of if checks in the update method.

TODO: add a 'resourceManager' and refactor the
base 'Engine'